+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/KagoMoshitateIdentifier.java
@@ -6,11 +6,10 @@
 package jp.co.ndensan.reams.db.dbc.business.core.basic;
 
 import java.io.Serializable;
-import lombok.Value;
 import jp.co.ndensan.reams.db.dbx.definition.core.valueobject.domain.HihokenshaNo;
 import jp.co.ndensan.reams.db.dbx.definition.core.valueobject.domain.JigyoshaNo;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleYearMonth;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
+import lombok.Value;
 
 /**
  * 過誤申立の識別子です。
@@ -21,7 +20,7 @@ public class KagoMoshitateIdentifier implements Serializable {
     private final JigyoshaNo 事業所番号;
     private final HihokenshaNo 被保険者番号;
     private final FlexibleYearMonth サービス提供年月;
-    private final Decimal 履歴番号;
+    private final int 履歴番号;
 
     /**
      * コンストラクタです。
@@ -34,7 +33,7 @@ public class KagoMoshitateIdentifier implements Serializable {
     public KagoMoshitateIdentifier(JigyoshaNo 事業所番号,
             HihokenshaNo 被保険者番号,
             FlexibleYearMonth サービス提供年月,
-            Decimal 履歴番号) {
+            int 履歴番号) {
         this.事業所番号 = 事業所番号;
         this.被保険者番号 = 被保険者番号;
         this.サービス提供年月 = サービス提供年月;
